Seguridad completa

+++ src/main/java/com/zeropokel/springprojects/tienda/repository/PermisoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.zeropokel.springprojects.tienda.model.Permiso;
 
 @Repository
-public interface PermisoRepository extends JpaRepository<Permiso, Long>{
+public interface PermisoRepository extends JpaRepository<Permiso, Integer>{
     
 }
